+++ ntex/tests/http_client.rs
@@ -1,9 +1,9 @@
 use std::io;
 
 use ntex::http::test::server as test_server;
-use ntex::http::{HttpService, Method, Request, Response, client::error::SendRequestError};
+use ntex::http::{HttpService, Method, Request, Response};
 use ntex::service::ServiceFactory;
-use ntex::{time, util::Bytes, util::Ready};
+use ntex::{client::error::SendRequestError, time, util::Bytes, util::Ready};
 
 const STR: &str = "Hello World Hello World Hello World Hello World Hello World \
                    Hello World Hello World Hello World Hello World Hello World \